% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imports_globalvars.R
\docType{package}
\name{TADAShinyAnalyze-package}
\alias{TADAShinyAnalyze}
\alias{TADAShinyAnalyze-package}
\title{TADAShinyAnalyze: TADAShinyAnalyze: Analyze Water and Sediment Chemistry Data}
\description{
The overall goal of the `TADAShinyAnalyze` shiny app is to provide a third module to the `TADAShiny` shiny app where users can complete batch and custom analyses by monitoring locations or assessment units (AUs) with the ultimate goal of supporting water quality management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USEPA/TADAShinyAnalyze}
  \item Report bugs at \url{https://github.com/USEPA/TADAShinyAnalyze/issues}
}

}
\author{
\strong{Maintainer}: Sheila Saia \email{sheila.saia@tetratech.com} (\href{https://orcid.org/0000-0002-4943-0306}{ORCID})

Authors:
\itemize{
  \item Cristina Mullin \email{mywaterway@epa.gov} (\href{https://orcid.org/0000-0002-0615-6087}{ORCID})
  \item Ben Block \email{ben.block@tetratech.com} (\href{https://orcid.org/0000-0003-1427-4363}{ORCID})
  \item Yu-Chen Wang
}

}
\keyword{internal}
